Cascade-main/requirements.txt
aiohttp==3.9.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
certifi==2024.6.2
charset-normalizer==3.3.2
Django==5.0.6
django-ranged-response==0.2.0
django-simple-captcha==0.6.0
frozenlist==1.4.1
idna==3.7
multidict==6.0.5
pillow==10.3.0
PyJWT==2.8.0
requests==2.32.3
sqlparse==0.5.0
twilio==9.2.3
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
yarl==1.9.4
